+++ example/custom_page_transitions_builder.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+
+class CustomPageTransitionsBuilder extends PageTransitionsBuilder {
+  @override
+  Widget buildTransitions<T>(
+    PageRoute<T> route,
+    BuildContext context,
+    Animation<double> animation,
+    Animation<double> secondaryAnimation,
+    Widget child,
+  ) {
+    return _CustomPageTransition(routeAnimation: animation, child: child);
+  }
+}
+
+class _CustomPageTransition extends StatelessWidget {
+  _CustomPageTransition({
+    Key? key,
+    required Animation<double> routeAnimation,
+    required this.child,
+  })  : _positionAnimation =
+            routeAnimation.drive(_bottomUpTween.chain(_fastOutSlowInTween)),
+        _opacityAnimation = routeAnimation.drive(_easeInTween),
+        super(key: key);
+
+  // Fractional offset from 1/4 screen below the top to fully on screen.
+  static final _bottomUpTween = Tween<Offset>(
+    begin: const Offset(0.0, 0.25),
+    end: Offset.zero,
+  );
+  static final Animatable<double> _fastOutSlowInTween =
+      CurveTween(curve: Curves.fastOutSlowIn);
+  static final Animatable<double> _easeInTween =
+      CurveTween(curve: Curves.easeIn);
+
+  final Animation<Offset> _positionAnimation;
+  final Animation<double> _opacityAnimation;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return SlideTransition(
+      position: _positionAnimation,
+      child: FadeTransition(opacity: _opacityAnimation, child: child),
+    );
+  }
+}