+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+* Add parameter performOutgoingAnimation for don't perform outgoing animation if the next route is a fullscreen dialog.
+
+## 1.0.1
+
+* Add function FlutterPageTransition.custom(RouteSettings, Widget, PageTransitionsBuilder) for custom page transition builder
+
 ## 1.0.0
 
 * Flutter custom page transitions builder